Added SKIP_CONFIRMATION variable to allow automation

+++ uninstall.sh
@@ -35,14 +35,16 @@ for f in $(find ./bin/ -mindepth 1 ); do
 done
 
 echo
-while true; do
-    read -p "Continue (y/n)?" yn
-    case $yn in
-        [Yy]* ) echo " "; break;;
-        [Nn]* ) exit 0;;
-        * ) echo "Please answer yes or no.";;
-    esac
-done
+if [[ "$SKIP_CONFIRMATION" != "true" ]]; then
+    while true; do
+        read -p "Continue (y/n)?" yn
+        case $yn in
+            [Yy]* ) echo " "; break;;
+            [Nn]* ) exit 0;;
+            * ) echo "Please answer yes or no.";;
+        esac
+    done
+fi
 
 sudo rm -vfr $FILESTOREMOVE
